set default make_safe_path=get_safe_path
